+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [CHANGE] Remove confirm prompt when closing office document #258
 * [FIX] Hide spinner only after all running requests have been completed #119
 * [CHANGE] Sort share consumers, allow scrolling #264
+* [CHANGE] Allow to remove a set destroy date #252
 
 ## 3.2.8
 **Maintainer**: balloon-team <opensource@gyselroth.net>\